+++ src/Task/GenerateViteConfig.php
@@ -11,9 +11,11 @@ namespace DecodeLabs\Zest\Task;
 use DecodeLabs\Atlas\File;
 use DecodeLabs\Clip\Task;
 use DecodeLabs\Clip\Task\GenerateFileTrait;
+use DecodeLabs\Coercion;
 use DecodeLabs\Overpass;
 use DecodeLabs\Terminus as Cli;
 use DecodeLabs\Zest;
+use DecodeLabs\Zest\Config;
 use DecodeLabs\Zest\Task\GenerateViteConfig\ViteTemplate;
 use DecodeLabs\Zest\Template;
 
@@ -21,14 +23,19 @@ class GenerateViteConfig implements Task
 {
     use GenerateFileTrait;
 
+    protected Config $config;
+
     protected function getTargetFile(): File
     {
         Cli::getCommandDefinition()
             ->addArgument('defaults=default', 'Defaults set name');
+        Cli::prepareArguments();
 
-        /** @var array<string, string> $args */
-        $args = Cli::prepareArguments();
-        Zest::$config->loadDefaults($args['defaults']);
+        $this->config = Zest::loadConfig();
+
+        $this->config->loadDefaults(
+            Coercion::toString(Cli::getArgument('defaults'))
+        );
 
         return Overpass::$rootDir->getFile('vite.config.js');
     }
@@ -36,13 +43,14 @@ class GenerateViteConfig implements Task
     protected function getTemplate(): ViteTemplate
     {
         return new ViteTemplate(
-            Zest::getController()
+            Zest::getController(),
+            $this->config
         );
     }
 
     protected function afterFileSave(File $file): bool
     {
-        Zest::$config->reload();
+        $this->config->reload();
 
         // Ensure index.html exists
         $index = Overpass::$rootDir->getFile('index.html');
@@ -57,7 +65,7 @@ class GenerateViteConfig implements Task
 
 
         // Ensure main.js exists
-        if (null !== ($entry = Zest::$config->getEntry())) {
+        if (null !== ($entry = $this->config->getEntry())) {
             $file = Zest::$package->rootDir->getFile($entry);
 
             if (!$file->exists()) {