+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Added multi-config setup support
+* Added custom manifest name support
+* Use npx directly to run vite
+
 ## v0.2.2 (2023-09-11)
 * Fixed dev manifest path issue